@jwt_required

+++ app/funcionario/controllers.py
@@ -2,7 +2,7 @@ from app.funcionario.model import Funcionario
 from flask import request, jsonify
 from app.extensions import db
 from flask.views import MethodView 
-from flask_jwt_extended import create_access_token, jwt_required
+from flask_jwt_extended import create_access_token, jwt_required, get_jwt_identity
 
 import bcrypt
 
@@ -43,12 +43,18 @@ class FuncionariosCreate (MethodView): #/funcionario/create
 
 class FuncionariosDetails(MethodView): #/funcionario/details/<int:id>
     decorators = [jwt_required()]
+    @jwt_required
     def get(self, id):
+        if (get_jwt_identity() != id):
+            return {'error':'Usuario não permitido'}, 400
+
         funcionario = Funcionario.query.get_or_404(id)
         
         return funcionario.json(), 200
 
     def put(self, id):
+        if (get_jwt_identity() != id):
+            return {'error':'Usuario não permitido'}, 400
         funcionario=Funcionario.query.get_or_404(id)
         dados = request.json
 
@@ -69,6 +75,10 @@ class FuncionariosDetails(MethodView): #/funcionario/details/<int:id>
         return funcionario.json(), 200
 
     def patch(self, id):
+        if (get_jwt_identity() != id):
+            return {'error':'Usuario não permitido'}, 400
+
+        funcionario = Funcionario.query.get_or_404(id)
         funcionario=Funcionario.query.get_or_404(id)
         dados = request.json
 
@@ -88,6 +98,8 @@ class FuncionariosDetails(MethodView): #/funcionario/details/<int:id>
         return funcionario.json(), 200
     
     def delete(self, id):
+        if (get_jwt_identity() != id):
+            return {'error':'Usuario não permitido'}, 400
         funcionario=Funcionario.query.get_or_404(id)
         db.session.delete(funcionario)
         db.session.commit()
